Fixed data packing 4

+++ labs/memory_bound/data_packing/solution.h
@@ -8,12 +8,12 @@ constexpr int maxRandom = 100;
 
 // FIXME: this data structure can be reduced in size
 struct S {
-	float	 d;
-	uint32_t l : 16;
-	uint32_t i : 8;
-	uint32_t s : 7;
-	bool	 b;
-	bool	 operator<(const S &s) const { return this->i < s.i; }
+	float	  d;
+	long long l : 16;
+	int		  i : 8;
+	short	  s : 7;
+	bool	  b;
+	bool	  operator<(const S &s) const { return this->i < s.i; }
 };
 
 template <int N>
